feat(UserQueryRepository): add method to retrieve all users and refactor package structure
