+++ backend/src/main/java/elbin_bank/issue_tracker/user/infrastructure/query/JdbcUserQueryRepository.java
@@ -1,7 +1,7 @@
 package elbin_bank.issue_tracker.user.infrastructure.query;
 
 import elbin_bank.issue_tracker.user.infrastructure.query.projection.UserProjection;
-import elbin_bank.issue_tracker.user.application.query.UserQueryRepository;
+import elbin_bank.issue_tracker.user.application.query.repository.UserQueryRepository;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
@@ -63,6 +63,25 @@ public class JdbcUserQueryRepository implements UserQueryRepository {
             }
             return result;
         });
+    }
+
+    @Override
+    public List<UserProjection> findAll() {
+        String sql = """
+                SELECT
+                  u.id,
+                  u.nickname,
+                  u.profile_image_url
+                FROM `user` u
+                WHERE u.deleted_at IS NULL
+                """;
+
+        return jdbc.query(sql, new MapSqlParameterSource(), (rs, rowNum) ->
+                new UserProjection(
+                        rs.getLong("id"),
+                        rs.getString("nickname"),
+                        rs.getString("profile_image_url")
+                ));
     }
 
 }
